slow printing idea
I added the delprint function to give it a scrolling text vibe. Might be too much of a pain though, because I'm not sure how to make it able to accept different numbers of arguments. On line 27 it reads it s three arguments rather than two. Not committed to this but let me know what you think.